+++ R/plots.R
@@ -1,14 +1,22 @@
 #' Plot Method for HAL Conditional Density Estimates
 #'
-#' @param x [TO FILL IN]
-#' @param ... [TO FILL IN]
-#' @param type [TO FILL IN]
+#' @param x Object of class \code{haldensify}, containing conditional density
+#'  estimates, as produced by \code{\link{haldensify}}.
+#' @param ... Additional arguments to be passed \code{plot}, currently ignored.
+#' @param type A \code{character} indicating the type of plot to be produced.
+#'  Options include visualizing the empirical risks of the conditional density
+#'  estimators across a grid of values of the regularization parameter and a
+#'  plot of the estimated conditional density (based on the estimator selected
+#'  by cross-validation). The latter has yet to be implemented.
 #'
-#' @importFrom data.table as.data.table
+#' @importFrom assertthat assert_that
+#' @importFrom data.table as.data.table data.table setnames
 #' @importFrom ggplot2 ggplot aes_string geom_point geom_line geom_vline xlab
 #'  ylab ggtitle theme_bw
+#' @importFrom latex2exp TeX
 #'
-#' @return [TO FILL IN]
+#' @return Object of class \code{ggplot} containing a plot of the desired
+#'  \code{type}.
 #'
 #' @export
 #'
@@ -27,24 +35,36 @@ plot.haldensify <- function(x, ..., type = c("risk", "density")) {
   # set default plot type
   type <- match.arg(type)
 
-  #
-  emp_risk_data <- data.table::as.data.table(list(
-    lambda = x$cv_tuning_results$lambda_seq,
-    risk = x$cv_tuning_results$emp_risks
-  ))
+  # density plot not yet implemented
+  assertthat::assert_that(type != "density",
+                          msg = "Density plot method not yet implemented.")
 
-  # create plot of empirical risks across grid in lambda
-  p_risk <- ggplot2::ggplot(
-      emp_risk_dat,
-      ggplot2::aes_string(x = lambda, y = risk)
-    ) +
-    ggplot2::geom_point() +
-    ggplot2::geom_line() +
-    ggplot2::geom_vline(xintercept = x$cv_tuning_results$lambda_loss_min,
-                        linetype = "dotted") +
-    ggplot2::xlab("Lambda (L1 regularization parameter)") +
-    ggplot2::ylab("Empirical Risk") +
-    ggplot2::ggtitle("Empirical risk of HAL conditional density estimators") +
-    ggplot2::theme_bw()
-  return(p_risk)
+  if (type == "risk") {
+    # construct 
+    emp_risk_data <- data.table::data.table(
+      lambda = x$cv_tuning_results$lambda_seq,
+      risk = x$cv_tuning_results$emp_risks
+    )
+    data.table::setnames(emp_risk_data, c("lambda", "risk"))
+    lambda_cvrisk_min <- x$cv_tuning_results$lambda_loss_min
+
+    # create plot of empirical risks across grid in lambda
+    p_emprisk <- ggplot2::ggplot(
+        emp_risk_data,
+        ggplot2::aes(x = lambda, y = risk)
+      ) +
+      ggplot2::geom_point() +
+      ggplot2::geom_line(linetype = "dashed") +
+      ggplot2::geom_vline(xintercept = lambda_cvrisk_min,
+                          linetype = "dotted") +
+      ggplot2::labs(
+        x = latex2exp::TeX("$L_1$ regularization parameter $\\lambda$"),
+        y = "Empirical risk",
+        title = "Empirical risk of HAL conditional density estimators",
+        subtitle = latex2exp::TeX(paste("(dotted line: choice of $\\lambda$",
+                                        "minimizing the empirical risk)"))
+      ) +
+      ggplot2::theme_bw()
+    return(p_emprisk)
+  }
 }